+++ lms/djangoapps/lti_provider/apps.py
@@ -13,4 +13,4 @@ class LtiProviderConfig(AppConfig):
 
     def ready(self):
         # Import the tasks module to ensure that signal handlers are registered.
-        from . import tasks  # pylint: disable=unused-import
+        from . import signals  # pylint: disable=unused-import
